Consolidate the two BeyondArenaContext classes into evaluation/context
PR #352 introduced nips2025_utils/beyond_arena_context.py without
noticing the existing evaluation/context/beyond_arena.py (added in
#339). Merge them into the existing one — the designated home for
pluggable evaluation contexts — keeping the union of both feature
sets:

- methods: new "beyond" preset (the Beyond-IID method collection) is
  now the default; "tabarena" and explicit lists still work.
- task_metadata: keeps the flexible resolution (suite name or
  *_tasks_metadata.csv path via load_beyond_task_metadata_collection,
  plus the native-schema DataFrame branch); "beyond" aliases
  "BeyondArena"; "tabarena" defers to the base loader.
- Preset resolution moves from inline __init__ string checks to the
  _resolve_*_preset hooks, with an explicit keyword-only signature.
- "lite" is re-keyed from fold == 0 to split == 0: on results frames
  "fold" holds the split index, so the original results-frame lambda
  meant split 0 == (fold 0, repeat 0); the fold-keyed grid predicate
  had silently diverged (fold 0 of every repeat). Now matches the base
  context's convention.

The duplicate class and its test file are removed; the unique test
coverage (preset resolution, grid/predicate compatibility, task-type
partitioning, split-keyed lite) moves to tst/evaluation/context/.
