+++ packages/tabarena/src/tabarena/evaluation/context/beyond_arena.py
@@ -2,20 +2,24 @@
 
 One of the pluggable evaluation contexts under :mod:`tabarena.evaluation.context`; subclasses
 :class:`~tabarena.nips2025_utils.tabarena_context.TabArenaContext`. BeyondArena differs from
-TabArena v0.1 in two ways:
+TabArena v0.1 in three ways:
 
 * **Subset predicates** — size buckets keyed on ``max_train_rows``, plus split-regime
-  (``random``/``temporal``/``grouped``), feature-dimensionality, text and high-cardinality subsets.
+  (``iid``/``temporal``/``grouped``), feature-dimensionality, text and high-cardinality subsets.
 * **Task metadata** — sourced from the self-contained committed BeyondArena reference CSV (via
   :func:`~tabarena.evaluation.beyond_metadata.load_beyond_task_metadata_collection`), whose tasks
   already carry the warehouse fields inline, so no separate ``warehouse_metadata.csv`` merge is
   needed.
+* **Method metadata** — the ``"beyond"`` preset selects the Beyond-IID benchmark's method
+  collection (artifact ``beyond_iid_benchmark_2026``; see
+  :mod:`tabarena.nips2025_utils.artifacts._beyond_method_metadata`).
 
 Everything else (method handling, plotting, leaderboard logic) is inherited unchanged.
 """
 
 from __future__ import annotations
 
+import copy
 from typing import TYPE_CHECKING
 
 import pandas as pd
@@ -25,6 +29,7 @@ from tabarena.nips2025_utils.tabarena_context import TabArenaContext
 
 if TYPE_CHECKING:
     from tabarena.benchmark.task.metadata import TaskMetadataCollection
+    from tabarena.models._method_metadata import MethodMetadata
 
 
 class BeyondArenaContext(TabArenaContext):
@@ -46,8 +51,9 @@ class BeyondArenaContext(TabArenaContext):
         "small": SubsetPredicate(lambda df: df["max_train_rows"].between(1_001, 10_000), ("max_train_rows",)),
         "tiny": SubsetPredicate(lambda df: df["max_train_rows"].between(101, 1_000), ("max_train_rows",)),
         # split / task type
-        "random": SubsetPredicate(lambda df: df["task_type"] == "random", ("task_type",)),  # IID; remove in future
         "iid": SubsetPredicate(lambda df: df["task_type"] == "random", ("task_type",)),
+        # backward-compatible alias of "iid"; remove in the future
+        "random": SubsetPredicate(lambda df: df["task_type"] == "random", ("task_type",)),
         "temporal": SubsetPredicate(lambda df: df["task_type"] == "temporal", ("task_type",)),
         "grouped": SubsetPredicate(lambda df: df["task_type"] == "grouped", ("task_type",)),
         # feature dimensionality / type
@@ -61,38 +67,41 @@ class BeyondArenaContext(TabArenaContext):
         "high-cardinality": SubsetPredicate(
             lambda df: df["num_high_cardinality_cats"] > 0, ("num_high_cardinality_cats",)
         ),
-        # row-level filter: keeps fold 0 of every repeat (the BeyondArena "lite" convention).
-        "lite": SubsetPredicate(lambda df: df["fold"] == 0, ("fold",)),
+        # split-level filter: keeps split 0 == (fold 0, repeat 0); a results frame's "fold"
+        # column is the split, so this maps to fold == 0 there (matching the original
+        # results-frame "lite" lambda and the base context's convention).
+        "lite": SubsetPredicate(lambda df: df["split"] == 0, ("split",)),
     }
 
     def __init__(
         self,
-        methods: str | list = "tabarena",
-        task_metadata: str | TaskMetadataCollection = "BeyondArena",
+        methods: list[MethodMetadata] | str = "beyond",
+        task_metadata: str | TaskMetadataCollection | pd.DataFrame = "BeyondArena",
         *,
+        extra_methods: list[MethodMetadata] | None = None,
+        backend: str = "ray",
         fillna_method: str | None = "RF (default)",
         calibration_method: str | None = "XGB (default)",
-        **kwargs,
     ) -> None:
         """Build a BeyondArena context.
 
         Args:
-            methods: Method preset/list, forwarded to :class:`TabArenaContext`.
-            task_metadata: A BeyondArena source name (e.g. ``"BeyondArena"``) or CSV path — loaded
-                via :func:`~tabarena.evaluation.beyond_metadata.load_beyond_task_metadata_collection`
-                — a ready ``TaskMetadataCollection``, or ``"tabarena"`` to defer to the base loader.
+            methods: ``"beyond"`` (the Beyond-IID method collection), ``"tabarena"`` (the
+                TabArena paper methods), or an explicit ``list[MethodMetadata]``.
+            task_metadata: A BeyondArena source name (e.g. ``"BeyondArena"``) or a
+                ``*_tasks_metadata.csv`` path — loaded via
+                :func:`~tabarena.evaluation.beyond_metadata.load_beyond_task_metadata_collection`
+                — a ready ``TaskMetadataCollection``, a native-schema (per-split) DataFrame,
+                or ``"tabarena"`` to defer to the base loader.
+            extra_methods: Additional ``MethodMetadata`` appended to the resolved methods.
+            backend: ``"ray"`` or ``"native"``, forwarded to :class:`TabArenaContext`.
             fillna_method: Imputed-method name forwarded to :class:`TabArenaContext`.
             calibration_method: Calibration-method name forwarded to :class:`TabArenaContext`.
-            **kwargs: Forwarded verbatim to :class:`TabArenaContext`.
         """
-        if isinstance(task_metadata, str) and task_metadata != "tabarena":
-            from tabarena.evaluation.beyond_metadata import load_beyond_task_metadata_collection
-
-            task_metadata = load_beyond_task_metadata_collection(task_metadata)
         if isinstance(task_metadata, pd.DataFrame):
             # The BeyondArena CSV is in the native (per-split) schema, not the legacy
             # one-row-per-dataset frame, so reconstruct a TaskMetadataCollection from it
-            # (matching `beyond_arena_eval`); the base context no longer accepts a DataFrame.
+            # (matching `beyond_arena_eval`); the base context does not accept a DataFrame.
             from tabarena.benchmark.task.metadata import InMemoryTaskMetadataSource, TaskMetadataCollection
 
             task_metadata = TaskMetadataCollection(InMemoryTaskMetadataSource(task_metadata).load())
@@ -100,10 +109,34 @@ class BeyondArenaContext(TabArenaContext):
         super().__init__(
             methods=methods,
             task_metadata=task_metadata,
+            extra_methods=extra_methods,
+            backend=backend,
             fillna_method=fillna_method,
             calibration_method=calibration_method,
-            **kwargs,
         )
+
+    def _resolve_task_metadata_preset(self, name: str) -> TaskMetadataCollection:
+        """``"tabarena"`` defers to the base loader; any other name is a BeyondArena suite
+        name (``"beyond"`` aliases ``"BeyondArena"``) or a ``*_tasks_metadata.csv`` path,
+        loaded self-contained (committed CSV, no downloads).
+        """
+        if name == "tabarena":
+            return super()._resolve_task_metadata_preset(name)
+        from tabarena.evaluation.beyond_metadata import load_beyond_task_metadata_collection
+
+        if name == "beyond":
+            name = "BeyondArena"
+        return load_beyond_task_metadata_collection(name)
+
+    def _resolve_methods_preset(self, name: str) -> list[MethodMetadata]:
+        """``"beyond"`` -> the Beyond-IID method collection; other names defer to
+        :class:`TabArenaContext`.
+        """
+        if name != "beyond":
+            return super()._resolve_methods_preset(name)
+        from tabarena.nips2025_utils.artifacts import beyond_method_metadata_collection
+
+        return copy.deepcopy(beyond_method_metadata_collection.method_metadata_lst)
 
     @property
     def _default_subsets(self):
